# assets/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "https://staging.spacebar.chat/api/",
            "description": "Official Spacebar Instance"
        }
    ],
    "info": {
        "description": "Spacebar is a free open source selfhostable discord compatible chat, voice and video platform",
        "version": "1.0.0",
        "title": "Spacebar HTTP API Routes",
        "termsOfService": "",
        "contact": {
            "name": "Spacebar"
        },
        "license": {
            "name": "AGPLV3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearer/Bot prefixes are not required."
            }
        },
...
    },
...
    "paths": {
...
        "/invites/{code}": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            },
            "post": {
                "x-right-required": "USE_MASS_INVITES",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            }
        },
...
    }
}